+++ src/main/java/com/cdkj/dzt/domain/Product.java
@@ -36,7 +36,7 @@ public class Product extends ABaseDO {
     private String description;
 
     // 损耗
-    private Integer loss;
+    private Double loss;
 
     // 加工费
     private Long processFee;
@@ -130,11 +130,11 @@ public class Product extends ABaseDO {
         this.description = description;
     }
 
-    public Integer getLoss() {
+    public Double getLoss() {
         return loss;
     }
 
-    public void setLoss(Integer loss) {
+    public void setLoss(Double loss) {
         this.loss = loss;
     }
